Refactored cover, decrypt, encrypt functions
