+++ pdf_bot/constants.py
@@ -7,7 +7,7 @@ WAIT_SCALE_BY_Y = 5
 WAIT_SCALE_TO_X = 6
 WAIT_SCALE_TO_Y = 7
 WAIT_SPLIT_RANGE = 8
-WAIT_FILENAME = 9
+WAIT_FILE_NAME = 9
 
 WAIT_FIRST_COMPARE_FILE = 0
 WAIT_SECOND_COMPARE_FILE = 1
